Feat: Payment 정산이 시작되었으면, OrderItem의 변경을 막는다.

+++ src/main/java/site/mylittlestore/form/ItemCreationForm.java
@@ -15,11 +15,11 @@ public class ItemCreationForm {
 
     @NotNull(message = "가격은 필수입니다.")
     @Min(value = 1, message = "가격은 0보다 커야합니다.")
-    private Integer price;
+    private Long price;
 
     @NotNull(message = "수량은 필수입니다.")
     @Min(value = 1, message = "수량은 0보다 커야합니다.")
-    private Integer stock;
+    private Long stock;
 
     private String image;
 }
